cluster_by_patristic_distance {HIVNetworkClustering}	R Documentation

Identify clusters in a phylogeny

Description

Uses a variation on the method of AFY Poon et al

Usage

cluster_by_patristic_distance(tree, min_size = 5, distance_threshold = 0.02)

Arguments

tree
A phylogenetic tree of the phylo class

min_size
Minimum size of components to report as clusters

distance_threshold
Pairwise patristic distance between taxa not exceeding this threshold will be clustered together

Details

As originally described, the adjacency matrix of the clustering graph is derived not from a single tree, but from 100 trees built from 100 bootstrap- sampled alignments. Patristic distance between taxa is calculated for each tree, and only those taxa below the threshold in a majority of trees have an edge in the final graph.

The default distance threshold of 0.02 substitutions/base is an empirically chosen value specific to the BC work by Poon et al., and is used here by way of example and not with any particular claim to quality (doi:10.1093/infdis/jiu560).

Value

A named numeric vector; the names are taxa labels from the tree, the values are the component of the clustering graph the named taxon is connected to, if the component size is at least min_size

[Package HIVNetworkClustering version 0.1 Index]