cluster_by_tn93_distance {HIVNetworkClustering}	R Documentation

Identify clusters in an alignment by nucleotide distance

Description

Uses TN93 distances with parameters from HIV-TRACE

Usage

cluster_by_tn93_distance(alignment, min_size = 5,
  distance_threshold = 0.015)

Arguments

alignment
A DNAStringSet of aligned sequences or a filesystem path to a multiple-sequence alignment

min_size
Minimum size of components to report as clusters

distance_threshold
Pairwise TN93 distance between taxa not exceeding this threshold will be clustered together

Value

A named numeric vector; the names are sequence ids from the alignment, the values are the component of the clustering graph the named taxon is connected to, if the component size is at least min_size

[Package HIVNetworkClustering version 0.1 Index]