graph_from_pairwise_distances {HIVNetworkClustering}	R Documentation

Construct a graph from a data frame of pairwise distances between nodes

Description

Edges are drawn between nodes within a specified distance of each other

Usage

graph_from_pairwise_distances(distances, max_distance)

Arguments

distances
A data frame where the first two columns contain node names and another column, distance, contains distances

max_distance
The distance above which edges are not drawn between nodes

Value

An undirected graph of class igraph

[Package HIVNetworkClustering version 0.1 Index]