tn93 {HIVNetworkClustering}	R Documentation

Calculate TN93 pairwise distances on a multiple-alignment

Description

Runs tn93 on a DNAStringSet or filesystem path.

Usage

tn93(alignment)

Arguments

dna_string_set
A Biostrings DNAStringSet of aligned sequences or a filesystem path to a multiple-sequence alignment

Details

Ambiguous base distances are minimized (resolved) between sequences with less than 1.5% ambiguous bases, and averaged between sequences with more than that. All pairwise distances are reported between sequences with at least 500 overlapping bases. These parameters are intended to match the CDC-recommended usage of HIV-TRACE.

The DNAStringSet or filesystem path must contain a multiple-sequence alignment.

The tn93 executable must be in your PATH.

Value

A data frame with three columns: id1, id2, and distance.

[Package HIVNetworkClustering version 0.1 Index]