tn93_distances {HIVNetworkClustering}	R Documentation

Compute the TN93 pairwise distances over an alignment

Description

Equivalent to tn93.

Usage

tn93_distances(alignment)

Arguments

alignment
A DNAStringSet of aligned sequences or a filesystem path to a multiple-sequence alignment

Value

A data frame of pairwise distances with three columns (id1, id2, and distance) suitable for passing to graph_from_pairwise_distances

[Package HIVNetworkClustering version 0.1 Index]